+++ api-editor/gui/src/app/App.tsx
@@ -70,7 +70,7 @@ import { AbstractPythonFilter } from '../features/filter/model/AbstractPythonFil
 import { UsageCountStore } from '../features/usages/model/UsageCountStore';
 import { PythonDeclaration } from '../features/packageData/model/PythonDeclaration';
 import { SaveFilterDialog } from '../features/filter/SaveFilterDialog';
-import { StatisticsView } from '../features/packageData/selectionView/StatisticsView';
+import { StatisticsView } from '../features/statistics/StatisticsView';
 
 export const App: React.FC = function () {
     useIndexedDB();
